# api-doc/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"CreateMenuDto": {
				"properties": {
					"icon": {
						"type": "string"
					},
					"menuName": {
						"type": "string"
					},
					"menuPath": {
						"type": "string"
					},
					"permNode": {
						"type": "number",
						"format": "double"
					},
					"level": {
						"type": "number",
						"format": "double"
					},
					"parentId": {
						"type": "number",
						"format": "double"
					},
					"permissionId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"menuName",
					"menuPath",
					"level",
					"parentId",
					"permissionId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateMenuDto": {
				"properties": {
					"icon": {
						"type": "string"
					},
					"menuName": {
						"type": "string"
					},
					"menuPath": {
						"type": "string"
					},
					"permNode": {
						"type": "number",
						"format": "double"
					},
					"level": {
						"type": "number",
						"format": "double"
					},
					"parentId": {
						"type": "number",
						"format": "double"
					},
					"permissionId": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"children": {
						"items": {
							"$ref": "#/components/schemas/UpdateMenuDto"
						},
						"type": "array"
					}
				},
				"required": [
					"menuName",
					"menuPath",
					"level",
					"parentId",
					"permissionId",
					"id",
					"children"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Required_Pick_GoodsTypes.typeParentId-or-typeEnable__": {
				"properties": {
					"typeEnable": {
						"type": "boolean"
					},
					"typeParentId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"typeEnable",
					"typeParentId"
				],
				"type": "object",
				"description": "Make all properties in T required"
			},
			"GoodsTypes": {
				"properties": {
					"goods": {
						"items": {
							"$ref": "#/components/schemas/Goods"
						},
						"type": "array"
					},
					"children": {
						"items": {
							"$ref": "#/components/schemas/GoodsTypes"
						},
						"type": "array"
					},
					"typeParent": {
						"$ref": "#/components/schemas/GoodsTypes"
					},
					"typeParentId": {
						"type": "number",
						"format": "double"
					},
					"typeEnable": {
						"type": "boolean"
					},
					"typeImg": {
						"type": "string"
					},
					"typeName": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"goods",
					"children",
					"typeEnable",
					"typeName",
					"id"
				],
				"type": "object"
			},
			"Comment": {
				"properties": {
					"goods": {
						"$ref": "#/components/schemas/Goods"
					},
					"children": {
						"items": {
							"$ref": "#/components/schemas/Comment"
						},
						"type": "array"
					},
					"parent": {
						"$ref": "#/components/schemas/Comment"
					},
					"authorId": {
						"type": "number",
						"format": "double"
					},
					"parentId": {
						"type": "number",
						"format": "double"
					},
					"content": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					},
					"goodsId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"goods",
					"children",
					"authorId",
					"content",
					"id",
					"goodsId"
				],
				"type": "object"
			},
			"Goods": {
				"properties": {
					"orderGoods": {
						"$ref": "#/components/schemas/OrderGoods"
					},
					"cartGoods": {
						"$ref": "#/components/schemas/CartGoods"
					},
					"goodsComments": {
						"items": {
							"$ref": "#/components/schemas/Comment"
						},
						"type": "array"
					},
					"goodsType": {
						"$ref": "#/components/schemas/GoodsTypes"
					},
					"goodsUserId": {
						"type": "number",
						"format": "double"
					},
					"goodsUser": {
						"type": "string"
					},
					"goodsUpdate": {
						"type": "string"
					},
					"goodsCreate": {
						"type": "string"
					},
					"goodsIsDel": {
						"type": "boolean"
					},
					"goodsSellCount": {
						"type": "number",
						"format": "double"
					},
					"goodsDesc": {
						"type": "string"
					},
					"goodsOnSale": {
						"type": "boolean"
					},
					"goodsTypeId": {
						"type": "number",
						"format": "double"
					},
					"goodsImgs": {
						"type": "string"
					},
					"goodsAmount": {
						"type": "number",
						"format": "double"
					},
					"goodsCostPrice": {
						"type": "number",
						"format": "double"
					},
					"goodsMarkPrice": {
						"type": "number",
						"format": "double"
					},
					"goodsPrice": {
						"type": "number",
						"format": "double"
					},
					"goodsName": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"goodsComments",
					"goodsType",
					"goodsUserId",
					"goodsUser",
					"goodsTypeId",
					"goodsImgs",
					"goodsCostPrice",
					"goodsMarkPrice",
					"goodsPrice",
					"goodsName",
					"id"
				],
				"type": "object"
			},
			"CartGoods": {
				"properties": {
					"goods": {
						"$ref": "#/components/schemas/Goods"
					},
					"cart": {
						"$ref": "#/components/schemas/Cart"
					},
					"goodsIsChecked": {
						"type": "boolean"
					},
					"goodsCount": {
						"type": "number",
						"format": "double"
					},
					"goodsId": {
						"type": "number",
						"format": "double"
					},
					"cartId": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"goods",
					"cart",
					"goodsIsChecked",
					"goodsCount",
					"goodsId",
					"cartId",
					"id"
				],
				"type": "object"
			},
			"Cart": {
				"properties": {
					"cartGoods": {
						"items": {
							"$ref": "#/components/schemas/CartGoods"
						},
						"type": "array"
					},
					"cartUserId": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"cartGoods",
					"cartUserId",
					"id"
				],
				"type": "object"
			},
			"OrderGoods": {
				"properties": {
					"order": {
						"$ref": "#/components/schemas/Orders"
					},
					"goods": {
						"$ref": "#/components/schemas/Goods"
					},
					"goodsCount": {
						"type": "number",
						"format": "double"
					},
					"goodsId": {
						"type": "number",
						"format": "double"
					},
					"orderId": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"order",
					"goods",
					"goodsCount",
					"goodsId",
					"orderId",
					"id"
				],
				"type": "object"
			},
			"Orders": {
				"properties": {
					"orderGoods": {
						"items": {
							"$ref": "#/components/schemas/OrderGoods"
						},
						"type": "array"
					},
					"orderUserId": {
						"type": "number",
						"format": "double"
					},
					"orderUpdate": {
						"type": "string"
					},
					"orderCreate": {
						"type": "string"
					},
					"orderStatus": {
						"type": "number",
						"format": "double"
					},
					"orderComment": {
						"type": "string"
					},
					"orderNo": {
						"type": "string"
					},
					"id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"orderGoods",
					"orderUserId",
					"orderStatus",
					"orderComment",
					"orderNo",
					"id"
				],
				"type": "object"
			},
			"Pick_GoodsTypes.Exclude_keyofGoodsTypes.id__": {
				"properties": {
					"typeName": {
						"type": "string"
					},
					"typeImg": {
						"type": "string"
					},
					"typeEnable": {
						"type": "boolean"
					},
					"typeParentId": {
						"type": "number",
						"format": "double"
					},
					"typeParent": {
						"$ref": "#/components/schemas/GoodsTypes"
					},
					"children": {
						"items": {
							"$ref": "#/components/schemas/GoodsTypes"
						},
						"type": "array"
					},
					"goods": {
						"items": {
							"$ref": "#/components/schemas/Goods"
						},
						"type": "array"
					}
				},
				"required": [
					"typeName",
					"typeEnable",
					"children",
					"goods"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_GoodsTypes.id_": {
				"$ref": "#/components/schemas/Pick_GoodsTypes.Exclude_keyofGoodsTypes.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Required_Pick_Goods.goodsName-or-goodsOnSale__": {
				"properties": {
					"goodsName": {
						"type": "string"
					},
					"goodsOnSale": {
						"type": "boolean"
					}
				},
				"required": [
					"goodsName",
					"goodsOnSale"
				],
				"type": "object",
				"description": "Make all properties in T required"
			},
			"Pick_Goods.Exclude_keyofGoods.id__": {
				"properties": {
					"goodsName": {
						"type": "string"
					},
					"goodsOnSale": {
						"type": "boolean"
					},
					"goodsPrice": {
						"type": "number",
						"format": "double"
					},
					"goodsMarkPrice": {
						"type": "number",
						"format": "double"
					},
					"goodsCostPrice": {
						"type": "number",
						"format": "double"
					},
					"goodsAmount": {
						"type": "number",
						"format": "double"
					},
					"goodsImgs": {
						"type": "string"
					},
					"goodsTypeId": {
						"type": "number",
						"format": "double"
					},
					"goodsDesc": {
						"type": "string"
					},
					"goodsSellCount": {
						"type": "number",
						"format": "double"
					},
					"goodsIsDel": {
						"type": "boolean"
					},
					"goodsCreate": {
						"type": "string"
					},
					"goodsUpdate": {
						"type": "string"
					},
					"goodsUser": {
						"type": "string"
					},
					"goodsUserId": {
						"type": "number",
						"format": "double"
					},
					"goodsType": {
						"$ref": "#/components/schemas/GoodsTypes"
					},
					"goodsComments": {
						"items": {
							"$ref": "#/components/schemas/Comment"
						},
						"type": "array"
					},
					"cartGoods": {
						"$ref": "#/components/schemas/CartGoods"
					},
					"orderGoods": {
						"$ref": "#/components/schemas/OrderGoods"
					}
				},
				"required": [
					"goodsName",
					"goodsPrice",
					"goodsMarkPrice",
					"goodsCostPrice",
					"goodsImgs",
					"goodsTypeId",
					"goodsUser",
					"goodsUserId",
					"goodsType",
					"goodsComments"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Goods.id_": {
				"$ref": "#/components/schemas/Pick_Goods.Exclude_keyofGoods.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "tsoa-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "LiuQi"
		}
	},
	"paths": {
...
		"/menu/icons": {
			"get": {
				"operationId": "GetIcons",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"label": {
												"type": "string"
											},
											"path": {
												"type": "string"
											}
										},
										"required": [
											"label",
											"path"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"菜单管理"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/menu/icon/upload": {
			"post": {
				"operationId": "UploadIcon",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"buffer": {
											"type": "string",
											"format": "byte"
										},
										"filename": {
											"type": "string"
										},
										"destination": {
											"type": "string"
										},
										"stream": {
											"type": "string",
											"format": "byte"
										},
										"size": {
											"type": "number",
											"format": "double"
										},
										"mimetype": {
											"type": "string"
										},
										"encoding": {
											"type": "string"
										},
										"originalname": {
											"type": "string"
										},
										"fieldname": {
											"type": "string"
										},
										"path": {}
									},
									"required": [
										"buffer",
										"filename",
										"destination",
										"stream",
										"size",
										"mimetype",
										"encoding",
										"originalname",
										"fieldname",
										"path"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"tags": [
					"菜单管理"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/menu/icon/delete": {
			"post": {
				"operationId": "DeleteIcon",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"菜单管理"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"filePath": {
										"type": "string"
									}
								},
								"required": [
									"filePath"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/menu/add": {
			"post": {
				"operationId": "AddMenuItem",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"菜单管理"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMenuDto"
							}
						}
					}
				}
			}
		},
		"/menu/list": {
			"post": {
				"operationId": "GetMenuList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CreateMenuDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"菜单管理"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMenuDto"
							}
						}
					}
				}
			}
		},
		"/menu/update": {
			"post": {
				"operationId": "UpdateMenuItem",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"菜单管理"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/CreateMenuDto"
									},
									{
										"properties": {
											"id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"id"
										],
										"type": "object"
									}
								]
							}
						}
					}
				}
			}
		},
		"/menu/updateSort": {
			"post": {
				"operationId": "UpdateMenuSort",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"菜单管理"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UpdateMenuDto"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/menu/delete": {
			"get": {
				"operationId": "DeleteMenuItem",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"菜单管理"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/goods-type/list": {
			"post": {
				"operationId": "GetTypeList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"data": {
											"items": {
												"properties": {
													"children": {
														"items": {
															"properties": {
																"typeParentId": {
																	"type": "number",
																	"format": "double"
																},
																"typeEnable": {
																	"type": "boolean"
																},
																"typeImg": {
																	"type": "string"
																},
																"typeName": {
																	"type": "string"
																},
																"id": {
																	"type": "number",
																	"format": "double"
																}
															},
															"required": [
																"typeParentId",
																"typeEnable",
																"typeImg",
																"typeName",
																"id"
															],
															"type": "object"
														},
														"type": "array"
													},
													"typeParentId": {
														"type": "number",
														"format": "double"
													},
													"typeEnable": {
														"type": "boolean"
													},
													"typeImg": {
														"type": "string"
													},
													"typeName": {
														"type": "string"
													},
													"id": {
														"type": "number",
														"format": "double"
													},
													"typeParentName": {
														"type": "string"
													}
												},
												"required": [
													"children",
													"typeParentId",
													"typeEnable",
													"typeImg",
													"typeName",
													"id",
													"typeParentName"
												],
												"type": "object"
											},
											"type": "array"
										},
										"code": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"message",
										"data",
										"code"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "获取商品类型列表",
				"tags": [
					"商品类型接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Required_Pick_GoodsTypes.typeParentId-or-typeEnable__"
							}
						}
					}
				}
			}
		},
		"/goods-type/add": {
			"post": {
				"operationId": "AddType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"data": {
											"properties": {
												"typeParentId": {
													"type": "number",
													"format": "double"
												},
												"typeEnable": {
													"type": "boolean"
												},
												"typeImg": {
													"type": "string"
												},
												"typeName": {
													"type": "string"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"typeParentId",
												"typeEnable",
												"typeImg",
												"typeName",
												"id"
											],
											"type": "object"
										},
										"code": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"message",
										"data",
										"code"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "添加商品类型",
				"tags": [
					"商品类型接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_GoodsTypes.id_"
							}
						}
					}
				}
			}
		},
		"/goods-type/update": {
			"post": {
				"operationId": "UpdateType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"data": {
											"properties": {
												"typeParentId": {
													"type": "number",
													"format": "double"
												},
												"typeEnable": {
													"type": "boolean"
												},
												"typeImg": {
													"type": "string"
												},
												"typeName": {
													"type": "string"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"typeParentId",
												"typeEnable",
												"typeImg",
												"typeName",
												"id"
											],
											"type": "object"
										},
										"code": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"message",
										"data",
										"code"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "更新商品类型",
				"tags": [
					"商品类型接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GoodsTypes"
							}
						}
					}
				}
			}
		},
		"/goods-type/delete": {
			"get": {
				"operationId": "DeleteType",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"data": {
											"properties": {
												"typeParentId": {
													"type": "number",
													"format": "double"
												},
												"typeEnable": {
													"type": "boolean"
												},
												"typeImg": {
													"type": "string"
												},
												"typeName": {
													"type": "string"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"typeParentId",
												"typeEnable",
												"typeImg",
												"typeName",
												"id"
											],
											"type": "object"
										},
										"code": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"message",
										"data",
										"code"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "删除商品类型",
				"tags": [
					"商品类型接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/goods/getList": {
			"post": {
				"operationId": "GetGoods",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"goodsUserId": {
												"type": "number",
												"format": "double"
											},
											"goodsUser": {
												"type": "string"
											},
											"goodsUpdate": {
												"type": "string",
												"format": "date-time"
											},
											"goodsCreate": {
												"type": "string",
												"format": "date-time"
											},
											"goodsIsDel": {
												"type": "boolean"
											},
											"goodsSellCount": {
												"type": "number",
												"format": "double"
											},
											"goodsDesc": {
												"type": "string"
											},
											"goodsOnSale": {
												"type": "boolean"
											},
											"goodsTypeId": {
												"type": "number",
												"format": "double"
											},
											"goodsAmount": {
												"type": "number",
												"format": "double"
											},
											"goodsCostPrice": {
												"type": "number",
												"format": "double"
											},
											"goodsMarkPrice": {
												"type": "number",
												"format": "double"
											},
											"goodsPrice": {
												"type": "number",
												"format": "double"
											},
											"goodsName": {
												"type": "string"
											},
											"id": {
												"type": "number",
												"format": "double"
											},
											"goodsImgs": {
												"items": {
													"type": "string"
												},
												"type": "array"
											}
										},
										"required": [
											"goodsUserId",
											"goodsUser",
											"goodsUpdate",
											"goodsCreate",
											"goodsIsDel",
											"goodsSellCount",
											"goodsDesc",
											"goodsOnSale",
											"goodsTypeId",
											"goodsAmount",
											"goodsCostPrice",
											"goodsMarkPrice",
											"goodsPrice",
											"goodsName",
											"id",
											"goodsImgs"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"商品接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Required_Pick_Goods.goodsName-or-goodsOnSale__"
							}
						}
					}
				}
			}
		},
		"/goods/create": {
			"post": {
				"operationId": "CreateGoods",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"商品接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/Omit_Goods.id_"
									},
									{
										"properties": {
											"goodsImgs": {
												"items": {
													"type": "string"
												},
												"type": "array"
											}
										},
										"required": [
											"goodsImgs"
										],
										"type": "object"
									}
								]
							}
						}
					}
				}
			}
		},
		"/goods/update": {
			"post": {
				"operationId": "UpdateGoods",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"商品接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/Goods"
									},
									{
										"properties": {
											"goodsImgs": {
												"items": {
													"type": "string"
												},
												"type": "array"
											}
										},
										"required": [
											"goodsImgs"
										],
										"type": "object"
									}
								]
							}
						}
					}
				}
			}
		},
		"/goods/delete": {
			"get": {
				"operationId": "DeleteGoods",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"商品接口"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}